Enhancement for Settings, when debounce is 0, we just commit changes for commit time Don´ call EEPROM get when no changes where found and brightness was 0 Fix for transition effect where the wrong data type was used Toggle key on remote now turns of current effect

+++ TransitionEffect.h
@@ -9,8 +9,8 @@
 class TransitionEffect final : public Effect {
 private:
     const HSB m_hsb;
-    const uint16_t m_startMillis;
-    const uint16_t m_endMillis;
+    const uint32_t m_startMillis;
+    const uint32_t m_endMillis;
     const uint16_t m_totalMillis;
 
 public:
